+++ rust/connlib/shared/src/messages/client.rs
@@ -215,30 +215,6 @@ impl ResourceDescription {
             }
         }
     }
-
-    fn is_internet_resource(&self) -> bool {
-        if let ResourceDescription::Internet(_) = self {
-            return true;
-        }
-
-        false
-    }
-}
-
-impl PartialOrd for ResourceDescription {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for ResourceDescription {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        if self.is_internet_resource() {
-            return std::cmp::Ordering::Less;
-        }
-
-        (self.name(), self.id()).cmp(&(other.name(), other.id()))
-    }
 }
 
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq, Hash)]
